LayoutTests/editing/selection/ios/show-grammar-replacements-on-tap-expected.txt

This test verifies that tapping on an editable grammar text checking range shows the edit menu and selects the range encompassing the grammatical error. This test requires WebKitTestRunner to run.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Presented edit menu
PASS getSelection().toString() is "thing"
PASS internals.hasGrammarMarker(0, 5) is true
PASS successfullyParsed is true

TEST COMPLETE
I'll need to thing about it.
